fix mousearea bug
This bug is due to use 'Component', when dynamicaly creating a Graph.qml (or anything else item with Mousearea). Previously used a StackView, which on button click created a menu component (AddWidgetPane.qml) on top of the stack. This menu component can dynamicaly create another item (in this case Graph.qml). But  after it's  poping (closing) from StackView, the MouseArea (including all buttons) of dynamicaly created item doesn't work. Maybe it's happens due to a violation of parental, i don't know.
Now the StackView is replaced by SwipeVIew,  OpenFilePane.qml and AddWidgetPane.qml is not a Components. It's imposible to use initialization by onCompleted for AddWidgetPane's ComboBoxes, and i am sending a tagList with a signal from c++ code.